# openapi.yaml
openapi: 3.0.3
info:
  title: 辰星认证中枢 API
  version: 1.0.0
  description: |
    辰星通行证用户认证、OAuth 2.0/OIDC 和统一用户管理 API。
    当前定义以仓库实际路由为准。运行期 Issuer 持久化在 PostgreSQL；未配置时仅健康检查、静态前端和初始化状态可用，其他协议路由由运行时门禁拒绝。

    请求超时（`REQUEST_TIMEOUT_SECONDS`）按路径分流错误信封（Issue #423）：
    `/oauth/*` 协议端点返回 RFC 6749 `503 temporarily_unavailable`（`OAuthErrorResponse`），
    不得改写为内部 API 信封；其余应用路由返回 `504` + `ErrorResponse`（`code=request_timeout`）。
    健康检查与静态 SPA fallback 不套用该请求超时。
servers:
  - url: http://127.0.0.1:3000
    description: 本地开发
...
paths:
...
  /.well-known/jwks.json:
    get:
      tags: [OAuth/OIDC]
      summary: JWKS 公钥集合
      operationId: jwks
      description: |
        返回当前及验证过渡期公钥集合，只用于验证 JWT，不包含私钥。
        公开只读元数据，CORS 与 Discovery 一致：请求带 Origin 时返回
        `Access-Control-Allow-Origin: *`（不带凭据），且始终 `Vary: Origin`。
        缓存策略为 `Cache-Control: public, max-age=60, must-revalidate`，RP 可在 60 秒内
        直接使用共享缓存；过期后用 `If-None-Match` 发起条件请求，公钥集合未变时返回 304。
        ETag 由 JWKS 字节的 SHA-256 派生，确定性且跨实例一致；密钥轮换或吊销后 ETag 改变。
      parameters:
        - { name: If-None-Match, in: header, required: false, description: RP 上次响应的 ETag，匹配时返回 304, schema: { type: string } }
      responses:
        '200':
          description: 公钥集合
          headers:
            Cache-Control:
              description: '`public, max-age=60, must-revalidate`'
              schema: { type: string }
            ETag:
              description: JWKS 字节 SHA-256 的强 ETag，用于条件请求。
              schema: { type: string }
            Access-Control-Allow-Origin:
              description: 当请求带 Origin 时返回 `*`，允许 RP 跨域读取；无 Origin 时省略。
              schema: { type: string }
            Vary:
              description: 始终为 `Origin`，避免共享缓存混用带/不带 Origin 的响应。
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
        '304':
          description: 公钥集合未变更，RP 可继续使用缓存副本
          headers:
            ETag:
              description: 与 200 响应相同的 ETag。
              schema: { type: string }
            Cache-Control:
              description: '`public, max-age=60, must-revalidate`'
              schema: { type: string }
            Access-Control-Allow-Origin:
              description: 当请求带 Origin 时返回 `*`；无 Origin 时省略。
              schema: { type: string }
            Vary:
              description: 始终为 `Origin`。
              schema: { type: string }
...
  /api/v1/auth/login:
    post:
      tags: [User Auth/Account]
      summary: 用户登录
      description: |
        错误码：
        - 401 invalid_credentials / invalid_factor：口令或 TOTP 验证码错误，或已被限流。
        - 503 factor_key_unavailable：请求带 `totp_code` 且该账号的 TOTP 种子由已退役的
          加密 kid 保护，服务端无法解密。重试无效，需要管理员重置因子。该结果不消耗
          认证失败额度。
      operationId: loginUser
      parameters:
        - { $ref: '#/components/parameters/SessionResponseMode' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: 登录成功，同时设置 Session 和 CSRF Cookie
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetCookie' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '202':
          description: 密码正确但需要首次绑定或完成认证因子。通过 HttpOnly login ticket 与 holder Cookie 继续流程，不在 JSON 响应中返回 ticket
          headers:
            Set-Cookie: { $ref: '#/components/headers/SetCookie' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PendingLoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /oauth/authorize:
    get:
      tags: [OAuth/OIDC]
      summary: OAuth 授权码入口
      operationId: authorize
      security: [{ sessionCookie: [] }, { chenxingSession: [] }, {}]
      description: |
        未登录浏览器请求仅在 client、redirect_uri、scope、state 和 PKCE 完整校验后创建 pending。
        X-Chenxing-Session 仅在 OAUTH_SESSION_HEADER_ENABLED=true 的开发兼容配置中接受，不能绕过 Session 绑定、CSRF 或 consent。
        请求超时在中间件层返回 503 `temporarily_unavailable`（RFC 6749 JSON 信封；此时尚无可用的 redirect_uri 上下文做错误重定向），不会返回内部 API 的 504 `request_timeout`（Issue #423）。
      parameters:
        - { $ref: '#/components/parameters/ClientId' }
        - { name: redirect_uri, in: query, required: true, schema: { type: string, format: uri } }
        - { name: response_type, in: query, required: true, schema: { type: string, enum: [code] } }
        - { name: scope, in: query, required: true, description: "空格分隔的 scope；每个值必须同时属于服务端 OAUTH_CLIENT_ALLOWED_SCOPES allowlist（默认 openid、profile、email）和该 Client 已注册的 scope 集合", schema: { type: string, example: 'openid email profile' } }
        - { name: state, in: query, required: true, description: 最多 512 个字符, schema: { type: string, maxLength: 512 } }
        - { name: nonce, in: query, description: 最多 512 个字符, schema: { type: string, maxLength: 512 } }
        - { name: code_challenge, in: query, required: true, schema: { type: string } }
        - { name: code_challenge_method, in: query, required: true, schema: { type: string, enum: [S256] } }
      responses:
        '302': { description: 重定向到 Client 的 redirect_uri 或登录/授权确认页（跳转到 SPA 登录页或授权确认页时附带授权持有者 HttpOnly Cookie；HTTPS 名称为 __Host-chenxing_authz_holder，本地 HTTP 名称为 chenxing_authz_holder，供后续绑定端点校验；已授权直通到 redirect_uri 的分支不下发） }
        '400': { $ref: '#/components/responses/OAuthBadRequest' }
        '401': { $ref: '#/components/responses/OAuthUnauthorized' }
        '429': { $ref: '#/components/responses/OAuthTooManyRequests' }
        '503': { $ref: '#/components/responses/OAuthServiceUnavailable' }
...
  /oauth/token:
    post:
      tags: [OAuth/OIDC]
      summary: 交换授权码或刷新 Token
      operationId: token
      description: |
        表单编码。Client 认证为 HTTP Basic 或 body 二选一。
        请求超时返回 503 `temporarily_unavailable`（RFC 6749 信封），不会返回内部 API 的 504 `request_timeout`（Issue #423）。
      security: [{ basicClientAuth: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Token 响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400': { $ref: '#/components/responses/OAuthBadRequest' }
        '401': { $ref: '#/components/responses/OAuthUnauthorized' }
        '429': { $ref: '#/components/responses/OAuthTooManyRequests' }
        '503': { $ref: '#/components/responses/OAuthServiceUnavailable' }
  /oauth/userinfo:
    get:
      tags: [OAuth/OIDC]
      summary: 获取 OIDC UserInfo
      operationId: userinfo
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401': { $ref: '#/components/responses/OAuthUnauthorized' }
        '503': { $ref: '#/components/responses/OAuthServiceUnavailable' }
  /oauth/revoke:
    post:
      tags: [OAuth/OIDC]
      summary: 撤销 Token
      operationId: revokeToken
      security: [{ basicClientAuth: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
                token_type_hint:
                  type: string
                  description: 可省略；仅支持 access_token 或 refresh_token，其他值返回 unsupported_token_type
                  enum: [access_token, refresh_token]
      responses:
        '200': { description: 撤销完成，无响应体 }
        '400': { $ref: '#/components/responses/OAuthBadRequest' }
        '401': { $ref: '#/components/responses/OAuthUnauthorized' }
        '503': { $ref: '#/components/responses/OAuthServiceUnavailable' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    adminBearer: { type: http, scheme: bearer, description: 系统 ADMIN_TOKEN 自动化通道，权限等价于 Owner 且无用户 ID，不需要浏览器 CSRF。ADMIN_TOKEN 为空时整个管理面关闭：Bearer 与 sessionCookie 两条通道都被拒绝，已初始化的管理接口统一返回 403 admin_disabled；唯一例外是不存在 Owner 时公开的首个 Owner 初始化接口 }
    basicClientAuth: { type: http, scheme: basic }
    sessionCookie: { type: apiKey, in: cookie, name: __Host-chenxing_session, description: HTTPS 浏览器部署使用的 host-only Session Cookie；仅 loopback HTTP 本地开发使用不带前缀的兼容名称 }
    chenxingSession: { type: apiKey, in: header, name: X-Chenxing-Session, description: 仅开发兼容配置 OAUTH_SESSION_HEADER_ENABLED=true 时可用；不能替代 pending、Session 绑定、CSRF 或 consent }
  parameters:
    ClientId: { name: client_id, in: query, required: true, schema: { type: string } }
    SessionResponseMode: { name: X-Chenxing-Session-Mode, in: header, required: false, description: 非浏览器兼容模式。仅在 SESSION_TOKEN_RESPONSE_ENABLED=true 且值为 token 时将 session_id 放入 JSON；浏览器不得发送此请求头。, schema: { type: string, enum: [token] } }
...
  headers:
    SetCookie: { description: Session、CSRF 或 pending-login Cookie, schema: { type: string } }
  responses:
...
    Unauthorized: { description: 未认证或凭据无效, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    ServiceUnavailable: { description: 依赖或必要能力暂不可用, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    OAuthBadRequest: { description: OAuth 请求错误，响应遵守 RFC 6749/7009, content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } } }
    OAuthUnauthorized:
      description: OAuth 客户端或 Bearer 凭据无效
      headers:
        WWW-Authenticate: { description: RFC 6749/6750 challenge, schema: { type: string } }
      content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } }
    OAuthTooManyRequests: { description: OAuth 限流或容量保护触发, content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } } }
    OAuthServiceUnavailable:
      description: |
        OAuth 依赖暂时不可用，或 `/oauth/*` 协议端点触发请求超时（Issue #423）。
        响应始终为 RFC 6749 信封（`error=temporarily_unavailable`），不是内部 `{code, message}`。
      content: { application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } } }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    OAuthErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string, description: RFC 6749/6750/7009 error code }
        error_description: { type: string, description: Human-readable non-sensitive description }
...
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier: { type: string, maxLength: 254, description: 用户名或邮箱；超过 254 字符按凭据错误拒绝 }
        password: { type: string, format: password, minLength: 10, maxLength: 128, description: 超过 128 字符按凭据错误拒绝，不返回独立错误码 }
        totp_code: { type: string, pattern: '^[0-9]{6}$' }
    PendingLoginResponse:
      type: object
      required: [status, methods]
      properties:
        status: { type: string, enum: [factor_setup_required, factor_required] }
        methods: { type: array, items: { type: string, enum: [totp, passkey] } }
...
    LoginResponse:
      type: object
      required: [expires_at]
      description: 浏览器安全的登录响应。Session token 默认不出现在 JSON 中，客户端应使用 HttpOnly Cookie 并通过 auth/status 和 auth/me 确认登录状态。
      properties:
        session_id: { type: string, description: '仅在 SESSION_TOKEN_RESPONSE_ENABLED=true 且请求显式发送 X-Chenxing-Session-Mode: token 时返回；浏览器响应默认不包含此字段' }
        expires_at: { type: string, format: date-time }
...
    TokenRequest:
      type: object
      required: [grant_type]
      properties:
        grant_type: { type: string, enum: [authorization_code, refresh_token] }
        code: { type: string }
        redirect_uri: { type: string, format: uri }
        code_verifier: { type: string }
        refresh_token: { type: string }
        scope:
          type: string
          description: refresh_token grant 可省略；空白字符串按省略处理并保留原授权 scope，非空值只能缩小原授权范围。
        client_id: { type: string }
        client_secret: { type: string, format: password }
    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, format: int64 }
        scope: { type: string }
        refresh_token: { type: string }
        id_token: { type: string, description: "JWT ID Token（scope 含 openid 时返回）" }
...
    UserInfo:
      type: object
      required: [sub]
      properties:
        sub: { type: string }
        email: { type: string, format: email }
        name: { type: string }
...